#1244:Added ImprovementType everywhere (DTOs, entities, seed data). Configured JSON to use JsonStringEnumConverter so enums accept "PGS" / "ISO". Fixed Swagger examples to show realistic values.

+++ backend/IMIS/IMIS.Application/IsoStandardModule/IIsoStandardService.cs
@@ -14,7 +14,7 @@ namespace IMIS.Application.IsoStandardModule
         Task<List<IsoStandardDto>?> FilterByClauseRef(string clauseRef, int noOfResults, CancellationToken cancellationToken);
         Task<DtoPageList<IsoStandardDto, IsoStandard, long>> GetPaginatedAsync(int page, int pageSize, CancellationToken cancellationToken);
         Task<DtoPageList<IsoStandardDto, IsoStandard, long>> GetPaginatedByVersionAsync(int versionId, int page, int pageSize, CancellationToken cancellationToken);
-        Task<bool> SoftDeleteAsync(long id, CancellationToken cancellationToken);
+        Task<bool> SoftDeleteAsync(int id, CancellationToken cancellationToken);
     }
 
 }
